requirements.txt
chainlit>=1.3.0
google-generativeai>=0.8.0
pandas>=2.0.0
